fix: Search Page Export & Refactor Search Library
