+++ src/lib/service/searchTypes.ts
