+++ src/lib/service/useSearchLogic.ts
